# Avoid apply in _compute_sum_of_squares for groupby std agg