--- a/dask/dataframe/groupby.py
+++ b/dask/dataframe/groupby.py
@@ -901,7 +901,7 @@ def _compute_sum_of_squares(grouped, column):
         keys = grouped.grouper
     else:
         # Handle CuDF groupby object (different from pandas)
-        keys = grouped.grouping._key_columns
+        keys = grouped.grouping.keys
     df = grouped.obj[column].pow(2) if column else grouped.obj.pow(2)
     return df.groupby(keys).sum()