@mrocklin 2020-05-08
Oh great!  I'm glad to see that it was possible to find a solution to this that used only more efficient Pandas/cudf API.  Seeing the speedup on pandas is great, and knowing that it allows cudf operations to work at all is awesome.  This seems to pass tests, so I'm +1, but I figure we wait a day in case someone like @TomAugspurger has thoughts.