@rjzamora 2020-05-08
Oh - It looks like Ashwin pointed out the `keys` attribute [here](https://github.com/dask/dask/issues/6034#issuecomment-625990025).